refactor: Rename `lib/plugins/aws/package/compile/events/httpApi.js`
