+++ lib/plugins/aws/package/compile/events/http-api.js
