+++ lib/plugins/index.js
@@ -52,7 +52,7 @@ module.exports = [
   require('./aws/package/compile/events/eventBridge/index.js'),
   require('./aws/package/compile/events/sqs.js'),
   require('./aws/package/compile/events/cloud-front.js'),
-  require('./aws/package/compile/events/httpApi'),
+  require('./aws/package/compile/events/http-api'),
   require('./aws/deploy-function.js'),
   require('./aws/deploy-list.js'),
   require('./aws/invoke-local/index.js'),
