+++ test/unit/lib/plugins/aws/package/compile/events/http-api.test.js
